fix(screen.css): code cut off on mobile devices
The existing approach relies on the `:hover` pseudo-class to make the code horizontally scrollable on small screens. But this doesn't work on mobile devices since you can't "hover" on a phone/tablet. Instead, the x-axis for code containers should reveal a scrollbar when appropriate. My PR fixes this by making the actual code example scrollable when the code overflows the container.